fix bug

+++ multitask_classifier.py
@@ -185,7 +185,7 @@ def train_multitask(args):
             b_ids_2 = b_ids_2.to(device)
             b_mask_2 = b_mask_2.to(device)
             logits = model.predict_paraphrase(b_ids_1, b_mask_1, b_ids_2, b_mask_2)
-            loss = F.mse_loss(logits, b_labels.view(-1), reduction='sum') / batch_size
+            loss = F.mse_loss(logits.float(), b_labels.view(-1), reduction='sum') / batch_size
         else:
             b_ids, b_mask, b_labels = (batch['token_ids'],
                                         batch['attention_mask'], batch['labels'])
@@ -194,7 +194,7 @@ def train_multitask(args):
             b_mask = b_mask.to(device)
             b_labels = b_labels.to(device)
             logits = model.predict_sentiment(b_ids, b_mask)
-            loss = F.cross_entropy(logits.float(), b_labels.view(-1), reduction='sum') / batch_size
+            loss = F.cross_entropy(logits, b_labels.view(-1), reduction='sum') / batch_size
         return loss
     
     # Run for the specified number of epochs
